refactor super resolution
